Commit text as-is when the cursor is moved.
Any "clicking around" cancels the current composing text. This is rather
restrictive. However, it is done this way currently as it seems impossible to
update the current composing text without jumping the cursor to / before the
start of the composing text or at / after the end. So if the user wanted to add a
ten to a kana somewhere in the middle of the composing text, there would be no
way to do this without making the cursor jump.
Thus, edits without committing are only allowed at the end of the composing text.
If this seems incorrect, please feel free to make a PR and justify your changes.